Move wallets endpoints to generic from admin. refactor imports (#82)
accordingly.
