+++ junit.xml
@@ -1,1 +1,5 @@
+<<<<<<< HEAD
 <?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="0" time="0.176" timestamp="2021-07-22T09:25:20.847625" hostname="localhost" /></testsuites>
+=======
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="0" time="0.195" timestamp="2021-07-26T11:28:25.135113" hostname="localhost" /></testsuites>
+>>>>>>> 80034fe... MOve wallets endpoints to generic from admin. refactor imports
